// src/mp4box.d.ts
export interface MP4VideoTrack extends MP4MediaTrack {
	video: MP4VideoData
}

export interface MP4AudioData {
	sample_rate: number
	channel_count: number
	sample_size: number
}

export interface MP4AudioTrack extends MP4MediaTrack {
	audio: MP4AudioData
}

export type MP4Track = MP4VideoTrack | MP4AudioTrack

export interface MP4Info {
	duration: number
	timescale: number
	fragment_duration: number
	isFragmented: boolean
	isProgressive: boolean
	hasIOD: boolean
	brands: string[]
	created: Date
	modified: Date
	tracks: MP4Track[]
	mime: string
	audioTracks: MP4AudioTrack[]
	videoTracks: MP4VideoTrack[]
}

export type MP4ArrayBuffer = ArrayBuffer & { fileStart: number }

export interface MP4File {
	onMoovStart?: () => void
	onReady?: (info: MP4Info) => void
	onError?: (e: string) => void
	onSamples?: (id: number, user: any, samples: Sample[]) => void

	appendBuffer(data: MP4ArrayBuffer): number
	start(): void
	stop(): void
	flush(): void

	setExtractionOptions(id: number, user: any, options: ExtractionOptions): void
}

export function createFile(): MP4File

export interface Sample {
	number: number
	track_id: number
	timescale: number
	description_index: number
	description: {
		avcC?: BoxParser.avcCBox // h.264
		hvcC?: BoxParser.hvcCBox // hevc
		vpcC?: BoxParser.vpcCBox // vp9
		av1C?: BoxParser.av1CBox // av1
	}
	data: ArrayBuffer
	size: number
	alreadyRead?: number
	duration: number
	cts: number
	dts: number
	is_sync: boolean
	is_leading?: number
	depends_on?: number
	is_depended_on?: number
	has_redundancy?: number
	degradation_priority?: number
	offset?: number
	subsamples?: any
}

export interface ExtractionOptions {
	nbSamples: number
}
// ...
export namespace BoxParser {
// ...
	export class av1CBox extends ContainerBox {
		constructor(size: any)

		parse(stream: any): void
	}
// ...
	export class avcCBox extends ContainerBox {
		constructor(size: any)

		parse(stream: any): void
		write(stream: any): void
	}
// ...
	export class hvcCBox extends ContainerBox {
		constructor(size: any)

		parse(stream: any): void
	}
// ...
	export class vpcCBox extends ContainerBox {
		constructor(size: any)

		parse(stream: any): void
	}
// ...
}
// ...
export namespace BoxParser {}